Rework how acl __setstate__ handles some errors
This resolves the very old comment 'Should we ignore errors'. Upon
more thinking, yes, ignoring errors is better, and the way the code
was, it contained a memory leak (on a very unlikely path).

So rework it to ignore errors in freeing the old ACL, since we don't
care about the old ACL, and the new one is successfully allocated.