Fix yet another acl_copy_int test that is causing segfaults
Version 0.7.0 broken in Debian on the s390x architecture, but not due
to the recently introduced tests, but a much older one. This didn't
break before because tests were broken on Debian (sigh again), which
hid the problem.

In order to make this in general more consistent, abstract the
"generate an invalid but safe(ish) external representation" and reuse
that in all the tests that cause, directly or indirectly, a call to
acl_copy_int().